email-api/requirements.txt
# Faculty Email Classification API
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database (aiosqlite for SQLite; asyncpg/psycopg2 only when using PostgreSQL)
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis & caching
redis==5.0.1
hiredis==2.3.2

# HTTP & WebSocket
httpx==0.26.0
python-multipart==0.0.9
websockets==12.0

# Utils
python-dotenv==1.0.1
email-validator>=2.0.0
python-dateutil==2.8.2
cryptography>=41.0.0

# Optional: Spark/Hive (for production with Hive)
# pyhive==0.7.0
# pyarrow==15.0.0
